requirements.txt
baldertest
balderplugin-junit
sphinx-rtd-theme>=1.2.0
sphinx==6.1.3
setuptools~=78.1.1
pycryptodome>=3.17
pylint==2.17.7
balderhub-unit
balderhub-data